Extract HomeScreen widget.
